deepcopy: gen: match Go 1.17 build tag format
The Go team has started the process of replacing the old "// +build"
constraints with the new "//go:build" constraints. See the design
doc for more details here:

https://go.googlesource.com/proposal/+/master/design/draft-gobuild.md

What is relevant is that in Go 1.17, gofmt now synchronizes the two
styles. In particular, if gofmt is run against a deepcopy generated
file, it will add a new "//go:build" constraint.

See https://golang.org/doc/go1.17#gofmt for details.

This PR intends to make it so that the generated code is compliant with
the new gofmt style.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>